move matcher -> select/string_matcher
