+++ src/select/sel_section.rs
@@ -1,6 +1,6 @@
-use crate::matcher::StringMatcher;
 use crate::md_elem::elem::*;
 use crate::select::match_selector::MatchSelector;
+use crate::select::string_matcher::StringMatcher;
 use crate::select::Matcher;
 
 #[derive(Debug, PartialEq)]
