+++ src/select/sel_single_matcher.rs
@@ -1,7 +1,7 @@
-use crate::matcher::StringMatcher;
 use crate::md_elem::elem::*;
 use crate::md_elem::elem_ref::*;
 use crate::select::match_selector::MatchSelector;
+use crate::select::string_matcher::StringMatcher;
 use crate::select::Matcher;
 use paste::paste;
